extract the mapping to an interface
