+++ logic/src/commonTest/kotlin/com/wire/kalium/logic/data/conversation/ConversationMapperTest.kt
@@ -156,87 +156,6 @@ class ConversationMapperTest {
         assertEquals(ConversationEntity.Type.GROUP, result.type)
     }
 
-    @Test
-    fun givenDomainAndTeamAreEqual_whenMappingToConversationDetails_ThenConversationDetailsUserTypeIsInternal() {
-        //given
-        val commonDomain = "commonDomain"
-        val commonTeam = "commonTeam"
-        val selfUser = generateTestSelfUser(commonDomain, commonTeam)
-        val otherUser = generateTestOtherUser(commonDomain, commonTeam)
-
-        given(userTypeMapper)
-            .function(userTypeMapper::fromOtherUserAndSelfUser)
-            .whenInvokedWith(any(), any())
-            .thenReturn(UserType.GUEST)
-        //when
-        val result = conversationMapper.toConversationDetailsOneToOne(TEST_CONVERSATION, otherUser, selfUser)
-        //then
-        assertEquals(UserType.INTERNAL, result.userType)
-    }
-
-    @Test
-    fun givenDomainAreDifferentButTeamsAreEqual_whenMappingToConversationDetails_ThenConversationDetailsUserTypeIsInternal() {
-        //given
-        val commonTeam = "commonTeam"
-
-        val selfUser = generateTestSelfUser("domainA", commonTeam)
-        val otherUser = generateTestOtherUser("domainB", commonTeam)
-
-        given(userTypeMapper)
-            .function(userTypeMapper::fromOtherUserAndSelfUser)
-            .whenInvokedWith(any(), any())
-            .thenReturn(UserType.GUEST)
-        //when
-        val result = conversationMapper.toConversationDetailsOneToOne(TEST_CONVERSATION, otherUser, selfUser)
-        //then
-        assertEquals(UserType.INTERNAL, result.userType)
-    }
-
-    @Test
-    fun givenCommonNotWireDomainAndDifferentTeam_whenMappingToConversationDetails_ThenConversationDetailsUserTypeIsFederated() {
-        //given
-        val commonDomain = "commonDomain"
-
-        val selfUser = generateTestSelfUser(commonDomain, "teamA")
-        val otherUser = generateTestOtherUser(commonDomain, "teamB")
-
-        given(userTypeMapper)
-            .function(userTypeMapper::fromOtherUserAndSelfUser)
-            .whenInvokedWith(any(), any())
-            .thenReturn(UserType.GUEST)
-        //when
-        val result = conversationMapper.toConversationDetailsOneToOne(TEST_CONVERSATION, otherUser, selfUser)
-        //then
-        assertEquals(UserType.FEDERATED, result.userType)
-    }
-
-    @Test
-    fun givenDifferentNotWireDomainAndDifferentTeam_whenMappingToConversationDetails_ThenConversationDetailsUserTypeIsFederated() {
-        //given
-        val selfUser = generateTestSelfUser("domainA", "teamA")
-        val otherUser = generateTestOtherUser("domainB", "teamB")
-
-        given(userTypeMapper)
-            .function(userTypeMapper::fromOtherUserAndSelfUser)
-            .whenInvokedWith(any(), any())
-            .thenReturn(UserType.GUEST)
-        //when
-        val result = conversationMapper.toConversationDetailsOneToOne(TEST_CONVERSATION, otherUser, selfUser)
-        //then
-        assertEquals(UserType.FEDERATED, result.userType)
-    }
-
-    @Test
-    fun givenUsingWireDomainAndDifferentTeam_whenMappingToConversationDetails_ThenConversationDetailsUserTypeIsGuest() {
-        //given
-        val selfUser = generateTestSelfUser("domainA", "teamA")
-        val otherUser = generateTestOtherUser("testDomain.wire.com", "teamB")
-        //when
-        val result = conversationMapper.toConversationDetailsOneToOne(TEST_CONVERSATION, otherUser, selfUser)
-        //then
-        assertEquals(UserType.GUEST, result.userType)
-    }
-
     private companion object {
         val ORIGINAL_CONVERSATION_ID = ConversationId("original", "oDomain")
         val SELF_USER_TEAM_ID = TeamId("teamID")
@@ -258,54 +177,5 @@ class ConversationMapperTest {
             ConvProtocol.PROTEUS,
             lastEventTime = "2022-03-30T15:36:00.000Z"
         )
-
-        val TEST_CONVERSATION = Conversation(
-            id = QualifiedID(
-                value = "testValue",
-                domain = "testDomain",
-            ),
-            name = "testName",
-            type = Conversation.Type.ONE_ON_ONE,
-            teamId = TeamId("test"),
-            mutedStatus = MutedConversationStatus.AllAllowed,
-            lastNotificationDate = "testNotificationDate",
-            lastModifiedDate = "testLastModifiedDate"
-        )
-
-        fun generateTestSelfUser(domain: String, team: String?): SelfUser {
-            return SelfUser(
-                id = QualifiedID(
-                    value = "testValue",
-                    domain = domain,
-                ),
-                name = "testName",
-                handle = "testHandle",
-                email = "testEmail",
-                phone = "testPhone",
-                accentId = 0,
-                team = team,
-                connectionStatus = ConnectionState.ACCEPTED,
-                previewPicture = "testPreviewPicture",
-                completePicture = "testCompletePicture"
-            )
-        }
-
-        fun generateTestOtherUser(domain: String, team: String?): OtherUser {
-            return OtherUser(
-                id = QualifiedID(
-                    value = "testValue",
-                    domain = domain,
-                ),
-                name = "testName",
-                handle = "testHandle",
-                email = "testEmail",
-                phone = "testPhone",
-                accentId = 0,
-                team = team,
-                connectionStatus = ConnectionState.ACCEPTED,
-                previewPicture = "testPreviewPicture",
-                completePicture = "testCompletePicture"
-            )
-        }
     }
 }
